Create PyNORXTESTS.py
Test case wrapper and executor.  You will need to "pip install colorama" and "pip install termcolor" to use the "cprint" method for color-coded results. This is by no means an exhaustive batch of tests, but allows for quick sanity check that everything is "working" as expected. REQUIRES PyNORXTESTCASES.py to work!